rename <navbar-bottom> as <quick-actions>
